Rename package name to homeschoolebt with refactor button
