--- src/main/java/org/formflowstartertemplate/app/StarterApplication.java
@@ -1,16 +1,0 @@
-package org.formflowstartertemplate.app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-
-@SpringBootApplication(scanBasePackages = {"org.formflowstartertemplate.app", "formflow.library"})
-@EnableConfigurationProperties
-public class StarterApplication {
-
-  public static void main(String[] args) {
-
-    SpringApplication.run(StarterApplication.class, args);
-  }
-
-}
