--- src/main/java/org/formflowstartertemplate/app/config/LocaleConfiguration.java
@@ -1,36 +1,0 @@
-package org.formflowstartertemplate.app.config;
-
-import java.util.Locale;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.ResourceBundleMessageSource;
-import org.springframework.web.servlet.LocaleResolver;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.i18n.CookieLocaleResolver;
-import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
-
-@Configuration
-public class LocaleConfiguration implements WebMvcConfigurer {
-
-  @Bean
-  public LocaleResolver localeResolver() {
-    CookieLocaleResolver localeResolver = new CookieLocaleResolver();
-    localeResolver.setCookieHttpOnly(true);
-    localeResolver.setCookieSecure(true);
-    localeResolver.setDefaultLocale(Locale.ENGLISH);
-    return localeResolver;
-  }
-
-  @Bean
-  public LocaleChangeInterceptor localeChangeInterceptor() {
-    LocaleChangeInterceptor interceptor = new LocaleChangeInterceptor();
-    interceptor.setParamName("lang");
-    return interceptor;
-  }
-
-  @Override
-  public void addInterceptors(InterceptorRegistry registry) {
-    registry.addInterceptor(localeChangeInterceptor());
-  }
-}
