--- src/main/java/org/formflowstartertemplate/app/config/SecurityConfiguration.java
@@ -1,19 +1,0 @@
-package org.formflowstartertemplate.app.config;
-
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfiguration {
-
-  @Bean
-  SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception {
-    httpSecurity.formLogin().disable();
-    return httpSecurity.build();
-  }
-}
