--- src/main/java/org/formflowstartertemplate/app/inputs/DocUpload.java
@@ -1,10 +1,0 @@
-package org.formflowstartertemplate.app.inputs;
-
-import formflow.library.data.FlowInputs;
-import lombok.Data;
-import org.springframework.web.multipart.MultipartFile;
-
-@Data
-public class DocUpload extends FlowInputs {
-  MultipartFile docUpload;
-}
