+++ src/main/java/org/homeschoolpebt/app/inputs/DocUpload.java
@@ -1,0 +1,10 @@
+package org.homeschoolpebt.app.inputs;
+
+import formflow.library.data.FlowInputs;
+import lombok.Data;
+import org.springframework.web.multipart.MultipartFile;
+
+@Data
+public class DocUpload extends FlowInputs {
+  MultipartFile docUpload;
+}
