+++ src/main/java/org/homeschoolpebt/app/submission/conditions/ExceedsIncomeThreshold.java
@@ -1,0 +1,22 @@
+package org.homeschoolpebt.app.submission.conditions;
+
+import formflow.library.config.submission.Condition;
+import formflow.library.data.Submission;
+import org.homeschoolpebt.app.utils.SubmissionUtilities;
+import org.springframework.stereotype.Component;
+
+/**
+ * This condition will check to see if the income the user has entered exceeds the maximum income allowable, based on their family
+ * size.
+ */
+@Component
+public class ExceedsIncomeThreshold implements Condition {
+
+  public Boolean run(Submission submission) {
+    double threshold = SubmissionUtilities.getIncomeThresholdByFamilySizeValue(submission);
+    double totalIncome = SubmissionUtilities.getHouseholdTotalIncomeValue(submission);
+
+    return totalIncome > threshold;
+  }
+
+}
