+++ src/main/java/org/homeschoolpebt/app/submission/conditions/HasHousehold.java
@@ -1,0 +1,17 @@
+package org.homeschoolpebt.app.submission.conditions;
+
+import formflow.library.config.submission.Condition;
+import formflow.library.data.Submission;
+import org.springframework.stereotype.Component;
+
+@Component
+public class HasHousehold implements Condition {
+
+  public Boolean run(Submission submission) {
+    var inputData = submission.getInputData();
+    if (inputData.containsKey("hasHousehold")) {
+      return inputData.get("hasHousehold").equals("true");
+    }
+    return false;
+  }
+}
