+++ src/main/resources/application.yaml
@@ -1,5 +1,5 @@
 form-flow:
-  inputs: 'org.formflowstartertemplate.app.inputs.'
+  inputs: 'org.homeschoolpebt.app.inputs.'
   uploads:
     accepted-file-types: '.jpeg,.jpg,.png,.pdf,.bmp,.gif,.doc,.docx,.odt,.ods,.odp'
     thumbnail-width: '64'
