+++ src/main/resources/templates/ubi/annualHouseholdIncome.html
@@ -17,7 +17,7 @@
             <strong th:text="#{annual-household-income.box-header}"></strong>
             <ul class="subflow-list list--bulleted padding-left-25">
               <th:block th:if="${inputData.containsKey('income')}">
-                <li th:each="personsIncome: ${T(org.formflowstartertemplate.app.utils.SubmissionUtilities).sortIncomeNamesWithApplicantFirst(submission)}"
+                <li th:each="personsIncome: ${T(org.homeschoolpebt.app.utils.SubmissionUtilities).sortIncomeNamesWithApplicantFirst(submission)}"
                     class="spacing-below-15">
                               <span class="grid">
                                 <span class="grid__item width-one-third">
@@ -27,7 +27,7 @@
                                 </span>
                                 <span
                                     class="grid__item width-one-third text--right padding-right-10"
-                                    th:with="personsTotalIncome=${T(org.formflowstartertemplate.app.utils.SubmissionUtilities).getIndividualsTotalIncome(submission, personsIncome.uuid)}"
+                                    th:with="personsTotalIncome=${T(org.homeschoolpebt.app.utils.SubmissionUtilities).getIndividualsTotalIncome(submission, personsIncome.uuid)}"
                                     th:text="${'$' + personsTotalIncome}"></span>
                                 <span
                                     class="grid__item width-one-third text--small spacing-below-0">
@@ -49,7 +49,7 @@
               <strong class="grid__item width-one-third text--left padding-left-25"
                       th:text="#{annual-household-income.total}"></strong>
               <strong class="grid__item width-one-third text--right"
-                      th:text="${'$' + T(org.formflowstartertemplate.app.utils.SubmissionUtilities).getHouseholdTotalIncome(submission)}"></strong>
+                      th:text="${'$' + T(org.homeschoolpebt.app.utils.SubmissionUtilities).getHouseholdTotalIncome(submission)}"></strong>
             </div>
             <a th:href="'/' + ${flow} + '/' + ${iterationStartScreen}"
                th:unless="${allHouseholdMembersHaveIncome}"
