--- src/test/java/org/formflowstartertemplate/app/utils/DatePart.java
@@ -1,15 +1,0 @@
-package org.formflowstartertemplate.app.utils;
-
-public enum DatePart {
-  YEAR(3), MONTH(1), DAY(2);
-
-  private final Integer position;
-
-  DatePart(Integer position) {
-    this.position = position;
-  }
-
-  public Integer getPosition() {
-    return this.position;
-  }
-}
