--- src/test/java/org/formflowstartertemplate/app/utils/YesNoAnswer.java
@@ -1,16 +1,0 @@
-package org.formflowstartertemplate.app.utils;
-
-public enum YesNoAnswer {
-  YES("Yes"),
-  NO("No");
-
-  private final String displayValue;
-
-  YesNoAnswer(String displayValue) {
-    this.displayValue = displayValue;
-  }
-
-  public String getDisplayValue() {
-    return displayValue;
-  }
-}
