+++ src/test/java/org/homeschoolpebt/app/utils/YesNoAnswer.java
@@ -1,0 +1,16 @@
+package org.homeschoolpebt.app.utils;
+
+public enum YesNoAnswer {
+  YES("Yes"),
+  NO("No");
+
+  private final String displayValue;
+
+  YesNoAnswer(String displayValue) {
+    this.displayValue = displayValue;
+  }
+
+  public String getDisplayValue() {
+    return displayValue;
+  }
+}
